Specify platform in integ test images
Otherwise, "make integration" will fail on M1 Macs. Note that this is a
workaround which assumes that, if "make integration" is run on an arm64
host, that host is able to run containers from amd64 images. This change
enables development on stereoscope on M1 Macs, but may not enable it on
other arm64 machines.

Signed-off-by: Will Murphy <will.murphy@anchore.com>