Update digests to use image ID
Previously, the test would loop over the RepoDigest field in the image
manifest to prove that the expect digest was present. However, in some
cases, the RepoDigests field can be empty. Therefore, just use the ID
from the manfiest, since this seems consisent across image sources, and
is always present.

Signed-off-by: Will Murphy <will.murphy@anchore.com>