chore: ungroup minor version dependabot updates
Dependabot make a lot of noise, so we've tried grouping all go updates
in this repo that are minor or patch. However, this makes a lot of pull
requests that fail unit tests and require human investigation, which is
harder than merging lots of tiny PRs that just turn green and get
merged.

Therefore, try only grouping go dependencies that are patch versions;
hopefully this will strike a balance between many PRs and PRs that don't
require work to get merged.

Signed-off-by: Will Murphy <willmurphyscode@users.noreply.github.com>