chore(deps): bump containerd to v2
This patch bumps containerd to v2. Containerd is a well behaved go
module, and so needs to be bumped explicitly in code since there are
some breaking changes (mostly code just being moved around). However,
AFAIK there are no wire-procotol changes.

This is needed since containerd v1 isn't really as actively updated. A
while back github.com/opencontainers/runtime-spec made a breaking change
in a minor update - the ecosystem is updating to the latest
runtime-spec. But containerd v1 uses the old version :( This means that
when a dependency updates to a new version of containerd v2 (which does
use the new version), then the old version can't also be included.

Super roundabout way of saying - if we move off of containerd v1 here,
then this helps here at least.

Before:

    $ go mod why github.com/containerd/containerd
    # github.com/containerd/containerd
    github.com/anchore/stereoscope/internal/containerd
    github.com/containerd/containerd

After:

    $ go mod why github.com/containerd/containerd
    # github.com/containerd/containerd
    (main module does not need package github.com/containerd/containerd)

Signed-off-by: Justin Chadwell <justin@unikraft.com>